+++ packages/@overeng/dotdot/pnpm-workspace.yaml.genie.ts
@@ -1,6 +1,0 @@
-import { commonPnpmWorkspaceData, pnpmWorkspaceYaml } from '../../../genie/internal.ts'
-
-export default pnpmWorkspaceYaml.manual({
-  packages: ['.'],
-  ...commonPnpmWorkspaceData,
-})