+++ packages/@overeng/megarepo/pnpm-workspace.yaml.genie.ts
@@ -1,8 +1,0 @@
-import { commonPnpmWorkspaceData, pnpmWorkspaceYaml } from '../../../genie/internal.ts'
-import pkg from './package.json.genie.ts'
-
-export default pnpmWorkspaceYaml.package({
-  pkg,
-  ...commonPnpmWorkspaceData,
-  publicHoistPattern: ['react', 'react-dom', 'react-reconciler'],
-})