+++ packages/@overeng/oxc-config/pnpm-workspace.yaml.genie.ts
@@ -1,7 +1,0 @@
-import { commonPnpmWorkspaceData, pnpmWorkspaceYaml } from '../../../genie/internal.ts'
-import pkg from './package.json.genie.ts'
-
-export default pnpmWorkspaceYaml.package({
-  pkg,
-  ...commonPnpmWorkspaceData,
-})